move Score class to separate file
